Get rid of warnings.
git-svn-id: https://svn.tdr.iem.uni-due.de:8443/svn/sctp/trunk/socketapi@163 516812f1-d3fd-0310-9ea4-9aededebcb87

+++ socket_programs/idexample1toM.c
@@ -89,7 +89,7 @@ my_recvmsg(int fd, struct msghdr *msg, void *buf, size_t *buflen, ssize_t *nrp, 
     }
 
     /* Realloc the buffer? */
-    if (*buflen == nr) {
+    if (*buflen == (size_t)nr) {
       buf = realloc(buf, *buflen * 2);
       if (buf == 0) {
         fprintf(stderr, "out of memory\n");
